requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# NLP Libraries
nltk>=3.8.0
spacy>=3.6.0
transformers>=4.30.0
torch>=2.0.0

# Feature Extraction
sentence-transformers>=2.2.0

# Deep Learning
tensorflow>=2.13.0  # Optional, for LSTM models

# Gradient Boosting
xgboost>=1.7.0

# API & Web
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0

# Database (Optional)
psycopg2-binary>=2.9.0
redis>=4.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Logging
loguru>=0.7.0